Domestication of the input sequence

The domesticate function checks for internal cleavage sites. If corresponding sites are present silent mutations are introduced to destroy the recognition sites. The functions returns a list containing the position of the choosen amino acid residue for silent mutation.

domesticate(
  input_sequence,
  restriction_enzyme = "GGTCTC",
  cuf = "e_coli_316407.csv"
)

Arguments

input_sequence
The sequence which should be modified. This is an object of type character containing the sequence.

restriction_enzyme
Recognition site sequence of the respective restriction enzyme [default: GGTCTC]

cuf
The Codon Usage Table which is being used to select the codon for an exchanged amino acid (and in this case to select the codon which shoulb be replaced). [default: e_coli_316407.csv]

Value

A list with replacments: Each element has a vector with the codon number at the first slot and the amino acid of this position at the second slot.

Examples

#Load the setup of the Point Mutation vignette and run the domestication
data(Point_Mutagenesis_BbsI_setup)
domesticate(input_sequence, restriction_enzyme=recognition_site_bbsi, cuf=cuf)

#> [[1]]
#> [1] "143" "K"  
#>