Calculate Primers for Multiple Site Saturation Mutagenesis

The mutate_msd function designs the necessary set of primers for the desired mutations.

msd_mutate(input_sequence, codon = "NDT", prefix = "TT",
  restriction_enzyme = "GGTCTC", suffix = "A", vector = c("AATG",
  "AAGC"), replacements, replacement_range = 5, binding_min_length = 4,
  binding_max_length = 9, target_temp = 60, fragment_min_size = 100)

Arguments

input_sequence
The sequence which should be modified. This is an object of type character containing the sequence.

codon
The desired type of MSD mutation [default: NDT]

prefix
Additional nucleobases in 5' position of the recognition site [default: TT]

restriction_enzyme
Recognition site sequence of the respective restriction enzyme [default: GGTCTC]

suffix
Spacer nucleotides matching the cleavage pattern of the enzyme [default: A]

vector
Four basepair overhangs complementary to the created overhangs in the acceptor vector [default: c("AATG", "AAGC")]

replacements
The desired substitutions

replacement_range
Maximum distance between two randomization sites to be incoporated into a single primer in amino acid residues [default: 5]

binding_min_length
The minimal threshold value of the length of the template binding sequence in amino acid residues [default: 4]

binding_max_length
Maximal length of the binding sequence in amino acid residues [default: 9]

target_temp
Melting temperature of the binding sequence in print('\u00B0')C [default: 60]

fragment_min_size
Minimal size of a generated gene fragment in base pairs [default 100]

Value

An object of class Primerset with the designed Primers.

Examples

#Load the setup of the MSD vignette and design the primers
data(MSD_BsaI_setup_lv2)
print(mutations)

#> [1] 137 143 147 232 234

print(recognition_site_bsai)

#> [1] "GGTCTC"

primers<-msd_mutate(input_sequence, prefix="TT" ,
restriction_enzyme=recognition_site_bsai, suffix="A",
vector=c("AATG", "AAGC"), replacements=mutations, replacement_range=5,
binding_min_length=4 , binding_max_length=9, target_temp=60,
fragment_min_size=60 )