Add a level to exisiting Primerset

This function replaces the prefix, the suffix and the restriction enzyme of a given Primerset to change the design to another level. You can use this function to convert an exisiting Level 2 Primerset to a Level 0 Primerset for example. Also the overhangs of the first and the last primer will be modified to match the plasmid of the new level.

primer_add_level(primerset, prefix = "TT",
  restriction_enzyme = "GAAGAC", suffix = "AA", vector = c("CTCA",
  "CTCG"))

Arguments

primerset
An exisiting Primerset (in Level 2)

prefix
Additional nucleobases in 5' position of the new recognition site [default: TT]

restriction_enzyme
Recognition site sequence of the new restriction enzyme (Level 0) [default: GAAGAC]

suffix
Spacer nucleotides matching the cleavage pattern of the enzyme (Level 0) [default: AA]

vector
Four basepair overhangs complementary to the created overhangs in the acceptor vector [default: c("CTCA", "CTCG")]

Value

A Primerset in the new Level (Level 0)

Examples

#Load level 2 results of the MSD vignette
data(MSD_BsaI_result_lv2)
primer_add_level(primers,  prefix="TT",
restriction_enzyme="GAAGAC", suffix="AA", vector=c("CTCA", "CTCG"))

#> An object of class "Extended_Primerset"
#> Slot "fragments":
#> [[1]]
#> An object of class "Fragment"
#> Slot "start":
#> [1] 2
#> 
#> Slot "stop":
#> [1] 142
#> 
#> Slot "start_mutation":
#> logical(0)
#> 
#> Slot "stop_mutation":
#> [1] 137
#> 
#> 
#> [[2]]
#> An object of class "Fragment"
#> Slot "start":
#> [1] 143
#> 
#> Slot "stop":
#> [1] 231
#> 
#> Slot "start_mutation":
#> [1] 143 147
#> 
#> Slot "stop_mutation":
#> logical(0)
#> 
#> 
#> [[3]]
#> An object of class "Fragment"
#> Slot "start":
#> [1] 232
#> 
#> Slot "stop":
#> [1] 300
#> 
#> Slot "start_mutation":
#> [1] 232 234
#> 
#> Slot "stop_mutation":
#> logical(0)
#> 
#> 
#> 
#> Slot "oldsequence":
#> [1] "ATGTCTCAGGTTCAGAGTGGCATTTTGCCAGAACATTGCCGCGCGGCGATTTGGATCGAAGCCAACGTGAAAGGGGAAGTTGACGCCCTGCGTGCGGCCAGTAAAACATTTGCCGACAAACTGGCAACTTTTGAAGCGAAATTCCCGGACGCGCATCTTGGTGCGGTGGTTGCCTTTGGTAACAACACCTGGCGCGCTCTGAGCGGCGGCGTTGGGGCAGAAGAGCTGAAAGATTTTCCGGGCTACGGTAAAGGCCTTGCGCCGACGACCCAGTTCGATGTGTTGATCCACATTCTTTCTCTGCGTCACGACGTAAACTTCTCTGTCGCCCAGGCGGCGATGGAAGCCTTTGGTGACTGCATTGAAGTGAAAGAAGAGATCCACGGCTTCCGTTGGGTTGAAGAGCGTGACCTGAGCGGCTTTGTTGACGGTACGGAAAACCCGGCGGGTGAAGAGACGCGTCGCGAAGTGGCGGTTATCAAAGACGGCGTGGATGCGGGCGGCAGCTATGTGTTTGTCCAGCGTTGGGAACACAACCTGAAGCAGCTCAACCGGATGAGCGTTCACGATCAGGAGATGGTGATCGGGCGCACCAAAGAGGCCAACGAAGAGATCGACGGCGACGAACGTCCGGAAACCTCTCACCTCACCCGCGTTGATCTGAAAGAAGATGGCAAAGGGCTGAAGATTGTTCGCCAGAGCCTGCCGTACGGCACTGCCAGTGGCACTCACGGTCTGTACTTCTGCGCCTACTGCGCGCGTCTGCATAACATTGAGCAGCAACTGCTGAGCATGTTTGGCGATACCGATGGTAAGCGTGATGCGATGTTGCGTTTCACCAAACCGGTAACCGGCGGCTATTATTTCGCACCGTCGCTGGACAAGTTGATGGCGCTGTAA"
#> 
#> Slot "primers":
#> [[1]]
#> [[1]][[1]]
#> An object of class "Primer_MSD"
#> Slot "prefix":
#> [1] "TT"
#> 
#> Slot "restriction_enzyme":
#> [1] "GAAGAC"
#> 
#> Slot "suffix":
#> [1] "AA"
#> 
#> Slot "vector":
#> [1] "CTCA"
#> 
#> Slot "overhang":
#> [1] "AATG"
#> 
#> Slot "extra":
#> character(0)
#> 
#> Slot "binding_sequence":
#> [1] "TCTCAGGTTCAGAGTGGCATTTTGCC"
#> 
#> Slot "temperature":
#> [1] 60.36446
#> 
#> Slot "difference":
#> [1] 0.3644563
#> 
#> 
#> [[1]][[2]]
#> An object of class "Primer_MSD"
#> Slot "prefix":
#> [1] "TT"
#> 
#> Slot "restriction_enzyme":
#> [1] "GAAGAC"
#> 
#> Slot "suffix":
#> [1] "AA"
#> 
#> Slot "vector":
#> [1] "CTCG"
#> 
#> Slot "overhang":
#> [1] "ACAA"
#> 
#> Slot "extra":
#> [1] "AGCCGCTCAGAHN"
#> 
#> Slot "binding_sequence":
#> [1] "ACGCTCTTCAACCCAACGGAAGC"
#> 
#> Slot "temperature":
#> [1] 59.785
#> 
#> Slot "difference":
#> [1] 0.5794564
#> 
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> An object of class "Primer_MSD"
#> Slot "prefix":
#> [1] "TT"
#> 
#> Slot "restriction_enzyme":
#> [1] "GAAGAC"
#> 
#> Slot "suffix":
#> [1] "AA"
#> 
#> Slot "vector":
#> [1] "CTCA"
#> 
#> Slot "overhang":
#> [1] "TTGT"
#> 
#> Slot "extra":
#> [1] "TNDTGGTACGGAANDT"
#> 
#> Slot "binding_sequence":
#> [1] "CCGGCGGGTGAAGAGACGC"
#> 
#> Slot "temperature":
#> [1] 59.2169
#> 
#> Slot "difference":
#> [1] 0.7831005
#> 
#> 
#> [[2]][[2]]
#> An object of class "Primer_MSD"
#> Slot "prefix":
#> [1] "TT"
#> 
#> Slot "restriction_enzyme":
#> [1] "GAAGAC"
#> 
#> Slot "suffix":
#> [1] "AA"
#> 
#> Slot "vector":
#> [1] "CTCG"
#> 
#> Slot "overhang":
#> [1] "AACA"
#> 
#> Slot "extra":
#> character(0)
#> 
#> Slot "binding_sequence":
#> [1] "ATCTTCAGCCCTTTGCCATCTTCTTT"
#> 
#> Slot "temperature":
#> [1] 59.04605
#> 
#> Slot "difference":
#> [1] 0.1708527
#> 
#> 
#> 
#> [[3]]
#> [[3]][[1]]
#> An object of class "Primer_MSD"
#> Slot "prefix":
#> [1] "TT"
#> 
#> Slot "restriction_enzyme":
#> [1] "GAAGAC"
#> 
#> Slot "suffix":
#> [1] "AA"
#> 
#> Slot "vector":
#> [1] "CTCA"
#> 
#> Slot "overhang":
#> [1] "TGTT"
#> 
#> Slot "extra":
#> [1] "NDTCAGNDT"
#> 
#> Slot "binding_sequence":
#> [1] "CTGCCGTACGGCACTGCCAG"
#> 
#> Slot "temperature":
#> [1] 60.52991
#> 
#> Slot "difference":
#> [1] 0.5299114
#> 
#> 
#> [[3]][[2]]
#> An object of class "Primer_MSD"
#> Slot "prefix":
#> [1] "TT"
#> 
#> Slot "restriction_enzyme":
#> [1] "GAAGAC"
#> 
#> Slot "suffix":
#> [1] "AA"
#> 
#> Slot "vector":
#> [1] "CTCG"
#> 
#> Slot "overhang":
#> [1] "AAGC"
#> 
#> Slot "extra":
#> character(0)
#> 
#> Slot "binding_sequence":
#> [1] "TTACAGCGCCATCAACTTGTCCAGC"
#> 
#> Slot "temperature":
#> [1] 60.90327
#> 
#> Slot "difference":
#> [1] 0.3733614
#> 
#> 
#> 
#> 
#> Slot "newsequence":
#> [1] "ATGTCTCAGGTTCAGAGTGGCATTTTGCCAGAACATTGCCGCGCGGCGATTTGGATCGAAGCCAACGTGAAAGGGGAAGTTGACGCCCTGCGTGCGGCCAGTAAAACATTTGCCGACAAACTGGCAACTTTTGAAGCGAAATTCCCGGACGCGCATCTTGGTGCGGTGGTTGCCTTTGGTAACAACACCTGGCGCGCTCTGAGCGGCGGCGTTGGGGCAGAAGAGCTGAAAGATTTTCCGGGCTACGGTAAAGGCCTTGCGCCGACGACCCAGTTCGATGTGTTGATCCACATTCTTTCTCTGCGTCACGACGTAAACTTCTCTGTCGCCCAGGCGGCGATGGAAGCCTTTGGTGACTGCATTGAAGTGAAAGAAGAGATCCACGGCTTCCGTTGGGTTGAAGAGCGTNDTCTGAGCGGCTTTGTTNDTGGTACGGAANDTCCGGCGGGTGAAGAGACGCGTCGCGAAGTGGCGGTTATCAAAGACGGCGTGGATGCGGGCGGCAGCTATGTGTTTGTCCAGCGTTGGGAACACAACCTGAAGCAGCTCAACCGGATGAGCGTTCACGATCAGGAGATGGTGATCGGGCGCACCAAAGAGGCCAACGAAGAGATCGACGGCGACGAACGTCCGGAAACCTCTCACCTCACCCGCGTTGATCTGAAAGAAGATGGCAAAGGGCTGAAGATTGTTNDTCAGNDTCTGCCGTACGGCACTGCCAGTGGCACTCACGGTCTGTACTTCTGCGCCTACTGCGCGCGTCTGCATAACATTGAGCAGCAACTGCTGAGCATGTTTGGCGATACCGATGGTAAGCGTGATGCGATGTTGCGTTTCACCAAACCGGTAACCGGCGGCTATTATTTCGCACCGTCGCTGGACAAGTTGATGGCGCTGTAA"
#>