Prepare a Single Point Mutation Primerset to be used in Level 2

This function adds definied vector overhangs to Level 0 Primersets to express them in a Level 2 vector. #'

primer_prepare_level(primerset, vector = c("AATG", "AAGC"))

Arguments

primerset
An exisiting Primerset (in Level 0)

vector
Four basepair overhangs complementary to the created overhangs in the acceptor vector [default: c("AATG", "AAGC")]

Value

A Primerset prepared for expression in Level 2

Examples

#Load level 0 results of the SPM vignette
data(Point_Mutagenesis_BbsI_result)
primer_prepare_level(primers)

#> An object of class "Extended_Primerset"
#> Slot "fragments":
#> [[1]]
#> An object of class "Fragment"
#> Slot "start":
#> [1] 1
#> 
#> Slot "stop":
#> [1] 66
#> 
#> Slot "start_mutation":
#> logical(0)
#> 
#> Slot "stop_mutation":
#> [1] 66
#> 
#> 
#> [[2]]
#> An object of class "Fragment"
#> Slot "start":
#> [1] 67
#> 
#> Slot "stop":
#> [1] 143
#> 
#> Slot "start_mutation":
#> logical(0)
#> 
#> Slot "stop_mutation":
#> [1] 143
#> 
#> 
#> [[3]]
#> An object of class "Fragment"
#> Slot "start":
#> [1] 144
#> 
#> Slot "stop":
#> [1] 246
#> 
#> Slot "start_mutation":
#> logical(0)
#> 
#> Slot "stop_mutation":
#> logical(0)
#> 
#> 
#> 
#> Slot "oldsequence":
#> [1] "ATGGTGAGCAAGGGCGAGGAGGATAACATGGCCATCATCAAGGAGTTCATGCGCTTCAAGGTGCACATGGAGGGCTCCGTGAACGGCCACGAGTTCGAGATCGAGGGCGAGGGCGAGGGCCGCCCCTACGAGGGCACCCAGACCGCCAAGCTGAAGGTGACCAAGGGTGGCCCCCTGCCCTTCGCCTGGGACATCCTGTCCCCTCAGTTCATGTACGGCTCCAAGGCCTACGTGAAGCACCCCGCCGACATCCCCGACTACTTGAAGCTGTCCTTCCCCGAGGGCTTCAAGTGGGAGCGCGTGATGAACTTCGAGGACGGCGGCGTGGTGACCGTGACCCAGGACTCCTCCCTGCAGGACGGCGAGTTCATCTACAAGGTGAAGCTGCGCGGCACCAACTTCCCCTCCGACGGCCCCGTAATGCAGAAGAAGACGATGGGCTGGGAGGCCTCCTCCGAGCGGATGTACCCCGAGGACGGCGCCCTGAAGGGCGAGATCAAGCAGAGGCTGAAGCTGAAGGACGGCGGCCACTACGACGCTGAGGTCAAGACCACCTACAAGGCCAAGAAGCCCGTGCAGCTGCCCGGCGCCTACAACGTCAACATCAAGTTGGACATCACCTCCCACAACGAGGACTACACCATCGTGGAACAGTACGAACGCGCCGAGGGCCGCCACTCCACCGGCGGCATGGACGAGCTGTACAAGGTCGACAAGCTTGCGGCCGCACTCGAGTGA"
#> 
#> Slot "primers":
#> [[1]]
#> [[1]][[1]]
#> An object of class "Primer_SPM"
#> Slot "prefix":
#> [1] "TT"
#> 
#> Slot "restriction_enzyme":
#> [1] "GAAGAC"
#> 
#> Slot "suffix":
#> [1] "AA"
#> 
#> Slot "vector":
#> [1] "CTCA"
#> 
#> Slot "overhang":
#> [1] ""
#> 
#> Slot "extra":
#> [1] "A"
#> 
#> Slot "binding_sequence":
#> [1] "ATGGTGAGCAAGGGCGAGGAGG"
#> 
#> Slot "temperature":
#> [1] 60.67248
#> 
#> Slot "difference":
#> [1] 0.6724812
#> 
#> 
#> [[1]][[2]]
#> An object of class "Primer_SPM"
#> Slot "prefix":
#> [1] "TT"
#> 
#> Slot "restriction_enzyme":
#> [1] "GAAGAC"
#> 
#> Slot "suffix":
#> [1] "TT"
#> 
#> Slot "vector":
#> [1] ""
#> 
#> Slot "overhang":
#> [1] "CACG"
#> 
#> Slot "extra":
#> [1] ""
#> 
#> Slot "binding_sequence":
#> [1] "ATGTCCCAGGCGAAGGGCAGGG"
#> 
#> Slot "temperature":
#> [1] 61.41454
#> 
#> Slot "difference":
#> [1] 0.7420589
#> 
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> An object of class "Primer_SPM"
#> Slot "prefix":
#> [1] "TT"
#> 
#> Slot "restriction_enzyme":
#> [1] "GAAGAC"
#> 
#> Slot "suffix":
#> [1] "AA"
#> 
#> Slot "vector":
#> [1] ""
#> 
#> Slot "overhang":
#> [1] "CGTG"
#> 
#> Slot "extra":
#> [1] ""
#> 
#> Slot "binding_sequence":
#> [1] "TCCCCTCAGTTCATGTACGGCTCC"
#> 
#> Slot "temperature":
#> [1] 59.45263
#> 
#> Slot "difference":
#> [1] 0.5473745
#> 
#> 
#> [[2]][[2]]
#> An object of class "Primer_SPM"
#> Slot "prefix":
#> [1] "TT"
#> 
#> Slot "restriction_enzyme":
#> [1] "GAAGAC"
#> 
#> Slot "suffix":
#> [1] "TT"
#> 
#> Slot "vector":
#> [1] ""
#> 
#> Slot "overhang":
#> [1] "TTTC"
#> 
#> Slot "extra":
#> [1] ""
#> 
#> Slot "binding_sequence":
#> [1] "TGCATTACGGGGCCGTCGGA"
#> 
#> Slot "temperature":
#> [1] 58.9345
#> 
#> Slot "difference":
#> [1] 0.5181218
#> 
#> 
#> 
#> [[3]]
#> [[3]][[1]]
#> An object of class "Primer_SPM"
#> Slot "prefix":
#> [1] "TT"
#> 
#> Slot "restriction_enzyme":
#> [1] "GAAGAC"
#> 
#> Slot "suffix":
#> [1] "AA"
#> 
#> Slot "vector":
#> [1] ""
#> 
#> Slot "overhang":
#> [1] "GAAA"
#> 
#> Slot "extra":
#> [1] ""
#> 
#> Slot "binding_sequence":
#> [1] "AAGACGATGGGCTGGGAGGCC"
#> 
#> Slot "temperature":
#> [1] 59.2929
#> 
#> Slot "difference":
#> [1] 0.7071014
#> 
#> 
#> [[3]][[2]]
#> An object of class "Primer_SPM"
#> Slot "prefix":
#> [1] "TT"
#> 
#> Slot "restriction_enzyme":
#> [1] "GAAGAC"
#> 
#> Slot "suffix":
#> [1] "TT"
#> 
#> Slot "vector":
#> [1] "CTCG"
#> 
#> Slot "overhang":
#> [1] ""
#> 
#> Slot "extra":
#> [1] "AAGC"
#> 
#> Slot "binding_sequence":
#> [1] "TCACTCGAGTGCGGCCGC"
#> 
#> Slot "temperature":
#> [1] 59.15354
#> 
#> Slot "difference":
#> [1] 0.1393632
#> 
#> 
#> 
#> 
#> Slot "newsequence":
#> [1] "ATGGTGAGCAAGGGCGAGGAGGATAACATGGCCATCATCAAGGAGTTCATGCGCTTCAAGGTGCACATGGAGGGCTCCGTGAACGGCCACGAGTTCGAGATCGAGGGCGAGGGCGAGGGCCGCCCCTACGAGGGCACCCAGACCGCCAAGCTGAAGGTGACCAAGGGTGGCCCCCTGCCCTTCGCCTGGGACATCGTGTCCCCTCAGTTCATGTACGGCTCCAAGGCCTACGTGAAGCACCCCGCCGACATCCCCGACTACTTGAAGCTGTCCTTCCCCGAGGGCTTCAAGTGGGAGCGCGTGATGAACTTCGAGGACGGCGGCGTGGTGACCGTGACCCAGGACTCCTCCCTGCAGGACGGCGAGTTCATCTACAAGGTGAAGCTGCGCGGCACCAACTTCCCCTCCGACGGCCCCGTAATGCAGAAAAAGACGATGGGCTGGGAGGCCTCCTCCGAGCGGATGTACCCCGAGGACGGCGCCCTGAAGGGCGAGATCAAGCAGAGGCTGAAGCTGAAGGACGGCGGCCACTACGACGCTGAGGTCAAGACCACCTACAAGGCCAAGAAGCCCGTGCAGCTGCCCGGCGCCTACAACGTCAACATCAAGTTGGACATCACCTCCCACAACGAGGACTACACCATCGTGGAACAGTACGAACGCGCCGAGGGCCGCCACTCCACCGGCGGCATGGACGAGCTGTACAAGGTCGACAAGCTTGCGGCCGCACTCGAGTGA"
#>