Add a level to exisiting Primerset

This function replaces the prefix, the suffix and the restriction enzyme of a given Primerset to change the design to another level. You can use this function to convert an exisiting Level 2 Primerset to a Level 0 Primerset for example. Also the overhangs of the first and the last primer will be modified to match the plasmid of the new level.

primer_add_level(
  primerset,
  prefix = "TT",
  restriction_enzyme = "GAAGAC",
  suffix = "AA",
  vector = c("CTCA", "CTCG")
)

Arguments

primerset
An exisiting Primerset (in Level 2)

prefix
Additional nucleobases in 5' position of the new recognition site [default: TT]

restriction_enzyme
Recognition site sequence of the new restriction enzyme (Level 0) [default: GAAGAC]

suffix
Spacer nucleotides matching the cleavage pattern of the enzyme (Level 0) [default: AA]

vector
Four basepair overhangs complementary to the created overhangs in the acceptor vector [default: c("CTCA", "CTCG")]

Value

A Primerset in the new Level (Level 0)

Examples

#Load level 2 results of the MSD vignette
data(MSD_BsaI_result_lv2)
primer_add_level(primers,  prefix="TT",
restriction_enzyme="GAAGAC", suffix="AA", vector=c("CTCA", "CTCG"))

#> An object of class "Extended_Primerset"
#> Slot "fragments":
#> [[1]]
#> An object of class "Fragment"
#> Slot "start":
#> [1] 2
#> 
#> Slot "stop":
#> [1] 141
#> 
#> Slot "start_mutation":
#> logical(0)
#> 
#> Slot "stop_mutation":
#> [1] 137
#> 
#> 
#> [[2]]
#> An object of class "Fragment"
#> Slot "start":
#> [1] 142
#> 
#> Slot "stop":
#> [1] 236
#> 
#> Slot "start_mutation":
#> [1] 143 147
#> 
#> Slot "stop_mutation":
#> [1] 232 234
#> 
#> 
#> [[3]]
#> An object of class "Fragment"
#> Slot "start":
#> [1] 237
#> 
#> Slot "stop":
#> [1] 300
#> 
#> Slot "start_mutation":
#> logical(0)
#> 
#> Slot "stop_mutation":
#> logical(0)
#> 
#> 
#> 
#> Slot "oldsequence":
#> [1] "ATGTCTCAGGTTCAGAGTGGCATTTTGCCAGAACATTGCCGCGCGGCGATTTGGATCGAAGCCAACGTGAAAGGGGAAGTTGACGCCCTGCGTGCGGCCAGTAAAACATTTGCCGACAAACTGGCAACTTTTGAAGCGAAATTCCCGGACGCGCATCTTGGTGCGGTGGTTGCCTTTGGTAACAACACCTGGCGCGCTCTGAGCGGCGGCGTTGGGGCAGAAGAGCTGAAAGATTTTCCGGGCTACGGTAAAGGCCTTGCGCCGACGACCCAGTTCGATGTGTTGATCCACATTCTTTCTCTGCGTCACGACGTAAACTTCTCTGTCGCCCAGGCGGCGATGGAAGCCTTTGGTGACTGCATTGAAGTGAAAGAAGAGATCCACGGCTTCCGTTGGGTTGAAGAGCGTGACCTGAGCGGCTTTGTTGACGGTACGGAAAACCCGGCGGGTGAAGAGACGCGTCGCGAAGTGGCGGTTATCAAAGACGGCGTGGATGCGGGCGGCAGCTATGTGTTTGTCCAGCGTTGGGAACACAACCTGAAGCAGCTCAACCGGATGAGCGTTCACGATCAGGAGATGGTGATCGGGCGCACCAAAGAGGCCAACGAAGAGATCGACGGCGACGAACGTCCGGAAACCTCTCACCTCACCCGCGTTGATCTGAAAGAAGATGGCAAAGGGCTGAAGATTGTTCGCCAGAGCCTGCCGTACGGCACTGCCAGTGGCACTCACGGTCTGTACTTCTGCGCCTACTGCGCGCGTCTGCATAACATTGAGCAGCAACTGCTGAGCATGTTTGGCGATACCGATGGTAAGCGTGATGCGATGTTGCGTTTCACCAAACCGGTAACCGGCGGCTATTATTTCGCACCGTCGCTGGACAAGTTGATGGCGCTGTAA"
#> 
#> Slot "primers":
#> [[1]]
#> [[1]][[1]]
#> An object of class "Primer_MSD"
#> Slot "prefix":
#> [1] "TT"
#> 
#> Slot "restriction_enzyme":
#> [1] "GAAGAC"
#> 
#> Slot "suffix":
#> [1] "AA"
#> 
#> Slot "vector":
#> [1] "CTCA"
#> 
#> Slot "overhang":
#> [1] "AATG"
#> 
#> Slot "extra":
#> [1] ""
#> 
#> Slot "binding_sequence":
#> [1] "TCTCAGGTTCAGAGTGGCATTTTGCC"
#> 
#> Slot "temperature":
#> [1] 60.36446
#> 
#> Slot "difference":
#> [1] 0.3644563
#> 
#> 
#> [[1]][[2]]
#> An object of class "Primer_MSD"
#> Slot "prefix":
#> [1] "TT"
#> 
#> Slot "restriction_enzyme":
#> [1] "GAAGAC"
#> 
#> Slot "suffix":
#> [1] "AA"
#> 
#> Slot "vector":
#> [1] "CTCG"
#> 
#> Slot "overhang":
#> [1] "AAAG"
#> 
#> Slot "extra":
#> [1] "CCGCTCAGAHN"
#> 
#> Slot "binding_sequence":
#> [1] "ACGCTCTTCAACCCAACGGAAGC"
#> 
#> Slot "temperature":
#> [1] 59.785
#> 
#> Slot "difference":
#> [1] 0.5794564
#> 
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> An object of class "Primer_MSD"
#> Slot "prefix":
#> [1] "TT"
#> 
#> Slot "restriction_enzyme":
#> [1] "GAAGAC"
#> 
#> Slot "suffix":
#> [1] "AA"
#> 
#> Slot "vector":
#> [1] "CTCA"
#> 
#> Slot "overhang":
#> [1] "CTTT"
#> 
#> Slot "extra":
#> [1] "GTTNDTGGTACGGAANDT"
#> 
#> Slot "binding_sequence":
#> [1] "CCGGCGGGTGAAGAGACGC"
#> 
#> Slot "temperature":
#> [1] 59.2169
#> 
#> Slot "difference":
#> [1] 0.7831005
#> 
#> 
#> [[2]][[2]]
#> An object of class "Primer_MSD"
#> Slot "prefix":
#> [1] "TT"
#> 
#> Slot "restriction_enzyme":
#> [1] "GAAGAC"
#> 
#> Slot "suffix":
#> [1] "AA"
#> 
#> Slot "vector":
#> [1] "CTCG"
#> 
#> Slot "overhang":
#> [1] "GGCA"
#> 
#> Slot "extra":
#> [1] "GAHNCTGAHN"
#> 
#> Slot "binding_sequence":
#> [1] "AACAATCTTCAGCCCTTTGCCATCTT"
#> 
#> Slot "temperature":
#> [1] 58.81557
#> 
#> Slot "difference":
#> [1] 0.401326
#> 
#> 
#> 
#> [[3]]
#> [[3]][[1]]
#> An object of class "Primer_MSD"
#> Slot "prefix":
#> [1] "TT"
#> 
#> Slot "restriction_enzyme":
#> [1] "GAAGAC"
#> 
#> Slot "suffix":
#> [1] "AA"
#> 
#> Slot "vector":
#> [1] "CTCA"
#> 
#> Slot "overhang":
#> [1] "TGCC"
#> 
#> Slot "extra":
#> [1] "G"
#> 
#> Slot "binding_sequence":
#> [1] "TACGGCACTGCCAGTGGCACT"
#> 
#> Slot "temperature":
#> [1] 59.7062
#> 
#> Slot "difference":
#> [1] 0.2937957
#> 
#> 
#> [[3]][[2]]
#> An object of class "Primer_MSD"
#> Slot "prefix":
#> [1] "TT"
#> 
#> Slot "restriction_enzyme":
#> [1] "GAAGAC"
#> 
#> Slot "suffix":
#> [1] "AA"
#> 
#> Slot "vector":
#> [1] "CTCG"
#> 
#> Slot "overhang":
#> [1] "AAGC"
#> 
#> Slot "extra":
#> [1] ""
#> 
#> Slot "binding_sequence":
#> [1] "TTACAGCGCCATCAACTTGTCCAG"
#> 
#> Slot "temperature":
#> [1] 58.868
#> 
#> Slot "difference":
#> [1] 0.8382065
#> 
#> 
#> 
#> 
#> Slot "newsequence":
#> [1] "ATGTCTCAGGTTCAGAGTGGCATTTTGCCAGAACATTGCCGCGCGGCGATTTGGATCGAAGCCAACGTGAAAGGGGAAGTTGACGCCCTGCGTGCGGCCAGTAAAACATTTGCCGACAAACTGGCAACTTTTGAAGCGAAATTCCCGGACGCGCATCTTGGTGCGGTGGTTGCCTTTGGTAACAACACCTGGCGCGCTCTGAGCGGCGGCGTTGGGGCAGAAGAGCTGAAAGATTTTCCGGGCTACGGTAAAGGCCTTGCGCCGACGACCCAGTTCGATGTGTTGATCCACATTCTTTCTCTGCGTCACGACGTAAACTTCTCTGTCGCCCAGGCGGCGATGGAAGCCTTTGGTGACTGCATTGAAGTGAAAGAAGAGATCCACGGCTTCCGTTGGGTTGAAGAGCGTNDTCTGAGCGGCTTTGTTNDTGGTACGGAANDTCCGGCGGGTGAAGAGACGCGTCGCGAAGTGGCGGTTATCAAAGACGGCGTGGATGCGGGCGGCAGCTATGTGTTTGTCCAGCGTTGGGAACACAACCTGAAGCAGCTCAACCGGATGAGCGTTCACGATCAGGAGATGGTGATCGGGCGCACCAAAGAGGCCAACGAAGAGATCGACGGCGACGAACGTCCGGAAACCTCTCACCTCACCCGCGTTGATCTGAAAGAAGATGGCAAAGGGCTGAAGATTGTTNDTCAGNDTCTGCCGTACGGCACTGCCAGTGGCACTCACGGTCTGTACTTCTGCGCCTACTGCGCGCGTCTGCATAACATTGAGCAGCAACTGCTGAGCATGTTTGGCGATACCGATGGTAAGCGTGATGCGATGTTGCGTTTCACCAAACCGGTAACCGGCGGCTATTATTTCGCACCGTCGCTGGACAAGTTGATGGCGCTGTAA"
#>